+++ db/migrate/20160809043433_add_status_description_to_bloom_remit_txns.rb
@@ -1,0 +1,5 @@
+class AddStatusDescriptionToBloomRemitTxns < ActiveRecord::Migration
+  def change
+    add_column :bloom_remit_txns, :status_description, :text
+  end
+end