+++ spec/dummy/db/schema.rb
@@ -11,24 +11,13 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20160808081300) do
+ActiveRecord::Schema.define(version: 20160809043433) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
-  enable_extension "hstore"
 
   create_table "billers", force: :cascade do |t|
     t.string "slug"
-  end
-
-  create_table "bloom_remit_billers", force: :cascade do |t|
-    t.string   "slug",                         null: false
-    t.text     "description"
-    t.hstore   "fields"
-    t.decimal  "service_charge", default: 0.0, null: false
-    t.datetime "deactivated_at"
-    t.datetime "created_at",                   null: false
-    t.datetime "updated_at",                   null: false
   end
 
   create_table "bloom_remit_responses", force: :cascade do |t|
@@ -41,17 +30,18 @@ ActiveRecord::Schema.define(version: 20160808081300) do
   add_index "bloom_remit_responses", ["txn_id"], name: "index_bloom_remit_responses_on_txn_id", using: :btree
 
   create_table "bloom_remit_txns", force: :cascade do |t|
-    t.integer "status",       default: 0, null: false
-    t.string  "sender_id",                null: false
-    t.string  "sender_type",              null: false
-    t.decimal "amount",                   null: false
+    t.integer "status",             default: 0, null: false
+    t.string  "sender_id",                      null: false
+    t.string  "sender_type",                    null: false
+    t.decimal "amount",                         null: false
     t.string  "account_name"
     t.string  "account_id"
-    t.uuid    "secret",                   null: false
-    t.string  "target_slug",              null: false
+    t.uuid    "secret",                         null: false
+    t.string  "target_slug",                    null: false
     t.string  "owner_id"
     t.string  "owner_type"
-    t.string  "external_id",              null: false
+    t.string  "external_id",                    null: false
+    t.text    "status_description"
   end
 
   add_index "bloom_remit_txns", ["external_id"], name: "index_bloom_remit_txns_on_external_id", using: :btree